PROTON-799: Wrap the pn_record_set and pn_record_clear Ruby APIs.
In wrapping these calls, we gain the ability to keep Ruby objects
alive that are effectively held only by the underlying C library. The
change adds an instance variable to the Swig wrapper to pn_record_t. This
instance variable, a hash, holds a key=value pair based on the key
and the value stores in the record. When the record is cleared, that
hash is also cleared.

And, most importantly, when the pn_record_t struct is discarded, all
of the underlying Ruby objects are also reaped along with it.